requirements.txt
joblib==1.2.0
keras==2.9.0
matplotlib==3.7.1
nibabel==5.1.0
nilearn==0.10.0
numpy==1.21.6
pandas==1.3.5
Pillow==9.4.0
Pillow==9.5.0
pycuda==2022.2.2
scikit_learn==1.0.2
scipy==1.10.1
tensorflow==2.9.0
torch==1.13.0
torchvision==0.14.0
tqdm==4.65.0
